# users-api/src/config/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Contact List API
# during dev, should point to your local machine
host: localhost:8000
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /users:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    get:
      description: Get all user of the system
      # used as the method name of the controller
      operationId: getAllUsers
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              # a pointer to a definition
              $ref: "#/definitions/BasicUser"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Update
      # used as the method name of the controller
      operationId: createUser
      parameters:
        - name: user
          in: body
          description: new user information
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/User"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/{id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    parameters:
      - name: id
        in: path
        description: the id of the user to retrieve
        required: true
        type: string
    get:
      description: Get a specific user by id
      # used as the method name of the controller
      operationId: getUser
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/User"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: Update
      # used as the method name of the controller
      operationId: editUser
      parameters:
        - name: user
          in: body
          description: new user information
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/User"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Update
      # used as the method name of the controller
      operationId: deleteUser
      responses:
        "204":
          description: Success
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  BasicUser:
    properties:
      _id:
        type: string
      gender:
        type: string
      name:
        type: object
        properties:
          title:
            type: string
          first:
            type: string
          last:
            type: string
      picture:
        type: object
        properties:
          thumbnail:
            type: string
          medium:
            type: string
          large:
            type: string
  User:
    properties:
      _id:
        type: string
      gender:
        type: string
      name:
        type: object
        properties:
          title:
            type: string
          first:
            type: string
          last: 
            type: string
      location:
        type: object
        properties:
          street:
            type: string
          city:
            type: string
          state:
            type: string
          zip:
            type: integer
      email:
        type: string
      username:
        type: string
      registered:
        type: integer
      dob:
        type: integer
      phone:
        type: string
      cell:
        type: string
      PPS:
        type: string
      picture:
        type: object
        properties:
          thumbnail:
            type: string
          medium:
            type: string
          large:
            type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string